Feature/melhorias

+++ src/pages/Theaters/Theaters.vue
@@ -1,44 +1,104 @@
 <template>
   <div class="card">
-    <VDialog header="Deletar Filme" :visible.sync="display">
-      <h6>inserir id</h6>
-      <input type="text" v-model="id" />
+    <VConfirmDialog> </VConfirmDialog>
+    <VDialog header="Limitar Lista" :visible.sync="display">
+      <span class="p-float-label mt-2 mb-4">
+        <VNumber type="text" v-model="limit" id="Quantidade" />
+        <label for="Quantidade">Quantidade</label>
+      </span>
+      <span class="p-float-label">
+        <VNumber type="text" v-model="skip" id="Pagina" />
+        <label for="Pagina">Pagina</label>
+      </span>
       <template #footer>
         <VButton
           label="Cancelar"
           icon="pi pi-times"
           class="p-button-text"
-          @click="hiddenDeleteDialog"
+          @click="hiddenPaginateDialog"
         />
         <VButton
-          label="Deletar"
+          label="Listar"
           icon="pi pi-check"
           autofocus
-          @click="deleteTheater"
+          @click="getTheatersPaginate"
         />
       </template>
     </VDialog>
-    <VDialog header="Criar Filme" :visible.sync="displayC">
-      <h6>TheaterId</h6>
-      <input type="text" v-model="Theaters.theaterId" />
-      <h5>Location:</h5>
-      <h6>Street1</h6>
-      <input type="text" v-model="Theaters.location.address.street1" />
-      <h6>Street2</h6>
-      <input type="text" v-model="Theaters.location.address.street2" />
-      <h6>Zipcode</h6>
-      <input type="text" v-model="Theaters.location.address.zipcode" />
-      <h6>State</h6>
-      <input type="text" v-model="Theaters.location.address.state" />
-      <h6>City</h6>
-      <input type="text" v-model="Theaters.location.address.city" />
-      <h5>Geo:</h5>
-      <h6>Type</h6>
-      <input type="text" v-model="Theaters.location.geo.type" />
-      <h6>Latitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[0]" />
-      <h6>Longitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[1]" />
+    <VDialog header="Criar Teatro" :visible.sync="displayC">
+      <div class="dialog mt-3">
+        <div class="inputField">
+          <span class="p-float-label">
+            <VNumber type="text" v-model="Theaters.theaterId" id="TheaterId" :useGrouping="false" />
+            <label for="TheaterId">TheaterId</label>
+          </span>
+          <span class="p-float-label">
+            <VInput
+              type="text"
+              v-model="Theaters.location.address.street1"
+              id="street1"
+            />
+            <label for="street1">street1</label>
+          </span>
+          <span class="p-float-label">
+            <VInput
+              type="text"
+              v-model="Theaters.location.address.street2"
+            />
+            <label for="street2">street2</label>
+          </span>
+        </div>
+        <div class="inputField">
+          <span class="p-float-label">
+            <VInput
+            type="text"
+            v-model="Theaters.location.address.zipcode"
+            id="zipcode"
+          />
+            <label for="zipcode">zipcode</label>
+          </span>
+          <span class="p-float-label">
+            <VInput
+            type="text"
+            v-model="Theaters.location.address.state"
+            id="state"
+          />
+            <label for="state">state</label>
+          </span>
+           <span class="p-float-label">
+            <VInput
+            type="text"
+            v-model="Theaters.location.address.city"
+            id="city"
+          />
+            <label for="city">city</label>
+          </span>
+        </div>
+        <div class="inputField">
+            <span class="p-float-label">
+             <VInput
+            type="text"
+            v-model="Theaters.location.geo.type"
+            id="type"
+          />
+            <label for="type">type</label>
+          </span>
+           <span class="p-float-label">
+             <VInput
+            v-model="Theaters.location.geo.coordinates[0]"
+            id="latitude"
+          />
+            <label for="latitude">latitude</label>
+          </span>
+           <span class="p-float-label">
+            <VInput
+            v-model="Theaters.location.geo.coordinates[1]"
+            id="longitude"
+          />
+            <label for="longitude">longitude</label>
+          </span>
+        </div>
+      </div>
       <template #footer>
         <VButton
           label="Cancelar"
@@ -54,29 +114,80 @@
         />
       </template>
     </VDialog>
-    <VDialog header="Atualizar Filme" :visible.sync="displayU">
-      <h6>inserir id</h6>
-      <input type="text" v-model="id" />
-      <h6>TheaterId</h6>
-      <input type="text" v-model="Theaters.location.address.theaterId" />
-      <h5>Location:</h5>
-      <h6>Street1</h6>
-      <input type="text" v-model="Theaters.location.address.street1" />
-      <h6>Street2</h6>
-      <input type="text" v-model="Theaters.location.address.street2" />
-      <h6>Zipcode</h6>
-      <input type="text" v-model="Theaters.location.address.zipcode" />
-      <h6>State</h6>
-      <input type="text" v-model="Theaters.location.address.state" />
-      <h6>City</h6>
-      <input type="text" v-model="Theaters.location.address.city" />
-      <h5>Geo:</h5>
-      <h6>Type</h6>
-      <input type="text" v-model="Theaters.location.geo.type" />
-      <h6>Latitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[0]" />
-      <h6>Longitude</h6>
-      <input type="text" v-model="Theaters.location.geo.coordinates[1]" />
+    <VDialog :header="updateMessage" :visible.sync="displayU">
+     <div class="dialog mt-3">
+        <div class="inputField">
+          <span class="p-float-label">
+            <VNumber type="text" v-model="Theaters.theaterId" id="TheaterId" :useGrouping="false" />
+            <label for="TheaterId">TheaterId</label>
+          </span>
+          <span class="p-float-label">
+            <VInput
+              type="text"
+              v-model="Theaters.location.address.street1"
+              id="street1"
+            />
+            <label for="street1">street1</label>
+          </span>
+          <span class="p-float-label">
+            <VInput
+              type="text"
+              v-model="Theaters.location.address.street2"
+            />
+            <label for="street2">street2</label>
+          </span>
+        </div>
+        <div class="inputField">
+          <span class="p-float-label">
+            <VInput
+            type="text"
+            v-model="Theaters.location.address.zipcode"
+            id="zipcode"
+          />
+            <label for="zipcode">zipcode</label>
+          </span>
+          <span class="p-float-label">
+            <VInput
+            type="text"
+            v-model="Theaters.location.address.state"
+            id="state"
+          />
+            <label for="state">state</label>
+          </span>
+           <span class="p-float-label">
+            <VInput
+            type="text"
+            v-model="Theaters.location.address.city"
+            id="city"
+          />
+            <label for="city">city</label>
+          </span>
+        </div>
+        <div class="inputField">
+            <span class="p-float-label">
+             <VInput
+            type="text"
+            v-model="Theaters.location.geo.type"
+            id="type"
+          />
+            <label for="type">type</label>
+          </span>
+           <span class="p-float-label">
+             <VInput
+            v-model="Theaters.location.geo.coordinates[0]"
+            id="latitude"
+          />
+            <label for="latitude">latitude</label>
+          </span>
+           <span class="p-float-label">
+            <VInput
+            v-model="Theaters.location.geo.coordinates[1]"
+            id="longitude"
+          />
+            <label for="longitude">longitude</label>
+          </span>
+        </div>
+      </div>
       <template #footer>
         <VButton
           label="Cancelar"
@@ -92,13 +203,19 @@
         />
       </template>
     </VDialog>
-    <div class="mr-4">
+    <div class="mx-2 my-3">
       <template>
         <VButton
-          label="Listar"
+          label="Listar Todos"
           icon="pi pi-list"
           class="p-button-primary mr-2"
           @click="showData"
+        />
+        <VButton
+          label="Limitar Listagem"
+          icon="pi pi-sliders-h"
+          class="p-button-info mr-2"
+          @click="showPaginateDialog"
         />
         <VButton
           label="Criar"
@@ -110,7 +227,7 @@
           label="Deletar"
           icon="pi pi-trash"
           class="p-button-danger mr-2"
-          @click="showDeleteDialog"
+          @click="deleteTheater"
         />
         <VButton
           label="Atualizar"
@@ -123,10 +240,11 @@
     <VDataTable
       :value="this.$store.state.Theaters"
       :paginator="true"
-      :rows="10"
-      stripedRows
+      :rows="15"
+      showGridlines
+      selectionMode="single"
+      @row-select="onRowSelect"
     >
-      <VColumn field="_id" header="ID"></VColumn>
       <VColumn field="theaterId" header="Theater-id"></VColumn>
       <VColumn field="location.address.street1" header="Street"></VColumn>
       <VColumn field="location.address.city" header="City"></VColumn>
@@ -137,16 +255,27 @@
 <script>
 export default {
   methods: {
-    showData() {
-      this.$store.dispatch(
-        "getTheaters",
-        `Bearer ${this.$store.state.jwtToken}`
-      );
+    onRowSelect(event) {
+      this.id = event.data._id;
+      this.TheaterName = event.data.theaterId;
+      this.updateMessage = `Atualizar Teatro: ${this.TheaterName}`;
     },
-    showDeleteDialog() {
+    showData() {
+      this.$store.dispatch("getTheaters", `Bearer ${this.jwt}`);
+    },
+    getTheatersPaginate() {
+      const data = {
+        jwt: `Bearer ${this.jwt}`,
+        limit: this.limit,
+        skip: this.skip,
+      };
+      this.$store.dispatch("getTheatersPaginate", data);
+      this.display = false;
+    },
+    showPaginateDialog() {
       this.display = true;
     },
-    hiddenDeleteDialog() {
+    hiddenPaginateDialog() {
       this.display = false;
     },
     showCreateDialog() {
@@ -162,31 +291,50 @@ export default {
       this.displayU = false;
     },
     deleteTheater() {
-      const data = {
-        id: this.id,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
-      };
-      this.$store.dispatch("deleteTheaters", data);
+      this.$confirm.require({
+        message: `Deseja Remover: ${this.TheaterName}`,
+        header: "Confirmação",
+        icon: "pi pi-exclamation-triangle",
+        acceptLabel: "Deletar",
+        rejectLabel: "Cancelar",
+        accept: () => {
+          const data = {
+            id: this.id,
+            jwt: `Bearer ${this.jwt}`,
+          };
+          this.$store.dispatch("deleteTheaters", data);
+        },
+      });
     },
     createTheater() {
-      this.Theaters.location.geo.coordinates[0] = parseFloat(this.Theaters.location.geo.coordinates[0]);
-      this.Theaters.location.geo.coordinates[1] = parseFloat(this.Theaters.location.geo.coordinates[1]);
+      this.Theaters.location.geo.coordinates[0] = parseFloat(
+        this.Theaters.location.geo.coordinates[0]
+      );
+      this.Theaters.location.geo.coordinates[1] = parseFloat(
+        this.Theaters.location.geo.coordinates[1]
+      );
       const data = {
         Theaters: this.Theaters,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
+        jwt: `Bearer ${this.jwt}`,
       };
       this.$store.dispatch("createTheaters", data);
+      this.displayC = false;
     },
     updateTheater() {
-      this.Theaters.location.geo.coordinates[0] = parseFloat(this.Theaters.location.geo.coordinates[0]);
-      this.Theaters.location.geo.coordinates[1] = parseFloat(this.Theaters.location.geo.coordinates[1]);
+      this.Theaters.location.geo.coordinates[0] = parseFloat(
+        this.Theaters.location.geo.coordinates[0]
+      );
+      this.Theaters.location.geo.coordinates[1] = parseFloat(
+        this.Theaters.location.geo.coordinates[1]
+      );
       const data = {
         Theaters: this.Theaters,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
+        jwt: `Bearer ${this.jwt}`,
         id: this.id,
       };
-      console.log(data)
+      console.log(data);
       this.$store.dispatch("updateTheaters", data);
+      this.displayU = false;
     },
   },
   data() {
@@ -194,9 +342,14 @@ export default {
       display: false,
       displayC: false,
       displayU: false,
+      jwt: localStorage.getItem("token"),
+      TheaterName: "",
+      updateMessage: "Nenhum teatro selecionado",
+      limit: undefined,
+      skip: undefined,
       id: "",
       Theaters: {
-        theaterId: "",
+        theaterId: undefined,
         location: {
           address: {
             street1: "",
@@ -216,5 +369,13 @@ export default {
 };
 </script>
 <style scoped>
+.dialog {
+  display: flex;
+  gap: 0.8rem;
+}
+.inputField{
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+}
 </style>
-
