+++ src/pages/Users/Users.vue
@@ -1,30 +1,45 @@
 <template>
   <div class="card">
-    <VDialog header="Deletar Usuario" :visible.sync="display">
-      <h6>inserir id</h6>
-      <input type="text" v-model="id" />
+    <VConfirmDialog> </VConfirmDialog>
+     <VDialog header="Limitar Lista" :visible.sync="display">
+      <span class="p-float-label mt-2 mb-4">
+        <VNumber type="text" v-model="limit" id="Quantidade" />
+        <label for="Quantidade">Quantidade</label>
+      </span>
+      <span class="p-float-label">
+        <VNumber type="text" v-model="skip" id="Pagina" />
+        <label for="Pagina">Pagina</label>
+      </span>
       <template #footer>
         <VButton
           label="Cancelar"
           icon="pi pi-times"
           class="p-button-text"
-          @click="hiddenDeleteDialog"
+          @click="hiddenPaginateDialog"
         />
         <VButton
-          label="Deletar"
+          label="Listar"
           icon="pi pi-check"
           autofocus
-          @click="deleteUser"
+          @click="getUsersPaginate"
         />
       </template>
     </VDialog>
     <VDialog header="Criar Usuario" :visible.sync="displayC">
-      <h6>Nome</h6>
-      <input type="text" v-model="Users.name" />
-      <h6>Email</h6>
-      <input type="text" v-model="Users.email" />
-      <h6>Senha</h6>
-      <input type="text" v-model="Users.password" />
+      <div class="dialog mt-2">
+        <span class="p-float-label">
+          <VInput type="text" v-model="Users.name" id="name" />
+          <label for="name">name</label>
+        </span>
+        <span class="p-float-label">
+          <VInput type="text" v-model="Users.email" id="email" />
+          <label for="email">email</label>
+        </span>
+        <span class="p-float-label">
+          <VInput type="text" v-model="Users.password" id="password" />
+          <label for="password">password</label>
+        </span>
+      </div>
       <template #footer>
         <VButton
           label="Cancelar"
@@ -39,16 +54,22 @@
           @click="createUser"
         />
       </template>
-    </VDialog>
-    <VDialog header="Atualizar Usuario" :visible.sync="displayU">
-      <h6>inserir id</h6>
-      <input type="text" v-model="id" />
-       <h6>Nome</h6>
-      <input type="text" v-model="Users.name" />
-      <h6>Email</h6>
-      <input type="text" v-model="Users.email" />
-      <h6>Senha</h6>
-      <input type="text" v-model="Users.password" />
+    </VDialog >
+    <VDialog :header="updateMessage" :visible.sync="displayU">
+      <div class="dialog mt-2">
+        <span class="p-float-label">
+          <VInput type="text" v-model="Users.name" id="name" />
+          <label for="name">name</label>
+        </span>
+        <span class="p-float-label">
+          <VInput type="text" v-model="Users.email" id="email" />
+          <label for="email">email</label>
+        </span>
+        <span class="p-float-label">
+          <VInput type="text" v-model="Users.password" id="password" />
+          <label for="password">password</label>
+        </span>
+      </div>
       <template #footer>
         <VButton
           label="Cancelar"
@@ -64,13 +85,19 @@
         />
       </template>
     </VDialog>
-    <div class="mr-4">
+    <div class="mx-2 my-3">
       <template>
         <VButton
-          label="Listar"
+          label="Listar Todos"
           icon="pi pi-list"
           class="p-button-primary mr-2"
           @click="showData"
+        />
+        <VButton
+          label="Limitar Listagem"
+          icon="pi pi-sliders-h"
+          class="p-button-info mr-2"
+          @click="showPaginateDialog"
         />
         <VButton
           label="Criar"
@@ -82,7 +109,7 @@
           label="Deletar"
           icon="pi pi-trash"
           class="p-button-danger mr-2"
-          @click="showDeleteDialog"
+          @click="deleteUser"
         />
         <VButton
           label="Atualizar"
@@ -95,10 +122,11 @@
     <VDataTable
       :value="this.$store.state.Users"
       :paginator="true"
-      :rows="10"
-      stripedRows
+      :rows="15"
+      showGridlines
+      selectionMode="single"
+      @row-select="onRowSelect"
     >
-      <VColumn field="_id" header="ID"></VColumn>
       <VColumn field="name" header="name"></VColumn>
       <VColumn field="email" header="email"></VColumn>
     </VDataTable>
@@ -107,13 +135,27 @@
 <script>
 export default {
   methods: {
-    showData() {
-      this.$store.dispatch("getUsers", `Bearer ${this.$store.state.jwtToken}`);
+    onRowSelect(event) {
+      this.id = event.data._id;
+      this.UserName = event.data.name;
+      this.updateMessage = `Atualizar Usuario: ${this.UserName}`;
     },
-    showDeleteDialog() {
+    getUsersPaginate() {
+      const data = {
+        jwt: `Bearer ${this.jwt}`,
+        limit: this.limit,
+        skip: this.skip,
+      };
+      this.$store.dispatch("getUsersPaginate", data);
+      this.display = false;
+    },
+    showData() {
+      this.$store.dispatch("getUsers", `Bearer ${this.jwt}`);
+    },
+    showPaginateDialog() {
       this.display = true;
     },
-    hiddenDeleteDialog() {
+    hiddenPaginateDialog() {
       this.display = false;
     },
     showCreateDialog() {
@@ -129,39 +171,63 @@ export default {
       this.displayU = false;
     },
     deleteUser() {
-      const data = {
-        id: this.id,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
-      };
-      this.$store.dispatch("deleteUsers", data);
+      this.$confirm.require({
+        message: `Deseja Remover: ${this.UserName}`,
+        header: "Confirmação",
+        icon: "pi pi-exclamation-triangle",
+        acceptLabel: "Deletar",
+        rejectLabel: "Cancelar",
+        accept: () => {
+          const data = {
+            id: this.id,
+            jwt: `Bearer ${this.jwt}`,
+          };
+          this.$store.dispatch("deleteUsers", data);
+        },
+      });
     },
     createUser() {
       this.$store.dispatch("createUsers", this.Users);
+      this.displayC = false;
     },
     updateUser() {
       const data = {
         Users: this.Users,
-        jwt: `Bearer ${this.$store.state.jwtToken}`,
-        id: this.id
+        jwt: `Bearer ${this.jwt}`,
+        id: this.id,
       };
-       this.$store.dispatch("updateUsers", data);
-    }
+      this.$store.dispatch("updateUsers", data);
+      this.displayU = false;
+    },
   },
   data() {
     return {
       display: false,
       displayC: false,
       displayU: false,
+      jwt: localStorage.getItem("token"),
+      UserName: "",
+      updateMessage: "Nenhum usuario selecionado",
+      limit: undefined,
+      skip: undefined,
       id: "",
       Users: {
-       name: "",
-       email: "",
-       password: ""
+        name: "",
+        email: "",
+        password: "",
       },
     };
   },
 };
 </script>
 <style scoped>
+.dialog {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  gap: 1.5rem;
+}
 </style>
